backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP client for API calls
httpx==0.25.2

# Environment variable management
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Optional: For advanced features
redis==5.0.1  # For caching
sqlalchemy==2.0.23  # For database if needed
alembic==1.12.1  # For database migrations if needed
